requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
arrow==1.3.0
attrs==24.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
dotmap==1.3.30
fonttools==4.53.1
glcontext==2.5.0
#glfw==2.7.0
imgui==2.0.0
kiwisolver==1.4.5
loguru==0.7.2
markdown-it-py==3.0.0
mdurl==0.1.2
moderngl==5.11.1
moderngl-window==2.4.6
multipledispatch==1.0.0
numpy==1.26.4
packaging==24.1
pillow==10.4.0
pydantic==2.6.1
pydantic_core==2.16.2
numpy-quaternion
Pygments==2.18.0
pyparsing==3.1.2
pyrr==0.10.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rich==13.7.1
shellingham==1.5.4
six==1.16.0
tqdm==4.66.5
turbopipe==1.2.1
typer==0.12.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
validators==0.33.0
watchdog==4.0.2
win32-setctime==1.1.0
